+++ benchmarking/commons/utils.py
@@ -39,12 +39,12 @@ sagemaker_estimator = {
 }
 
 
-# Used for simulator back-end experiments and for remote launching of
-# SageMaker back-end experiments
+# Used for simulator backend experiments and for remote launching of
+# SageMaker backend experiments
 def basic_cpu_instance_sagemaker_estimator(**kwargs):
     """
-    Returns SageMaker estimator to be used for simulator back-end experiments
-    and for remote launching of SageMaker back-end experiments.
+    Returns SageMaker estimator to be used for simulator backend experiments
+    and for remote launching of SageMaker backend experiments.
 
     :param kwargs: Extra arguments to SageMaker estimator
     :return: SageMaker estimator