+++ tst/test_constrained_bo.py
@@ -20,7 +20,7 @@ from syne_tune.config_space import uniform
 
 
 # TODO: Turn ``train_constrained_example`` into a blackbox and use the simulator
-# back-end to run this test rapidly
+# backend to run this test rapidly
 @pytest.mark.skip("this unit test takes about two minutes and is skipped for now")
 @pytest.mark.parametrize(
     "scheduler, searcher, constraint_offset",